Update README.md
I updated:
Schedule an appointment: to use a 6 digit number as MRN. This MRN will be unique per patient and should tie to the correct patient.

Cancel an appointment: to use an appointment_id. This may make querying and deleting a unique appointment a bit easier.

Open to other suggestions.